chore: move models into folders and namespaces
