+++ src/Explore.Cli/Models/ExploreCliModels.cs
@@ -1,3 +1,5 @@
+namespace Explore.Cli.Models;
+
 public partial class SchemaValidationResult {
     public bool isValid { get; set; } = false;
     public string? Message { get; set; }
